Extracts AbstractClockerIntegrationTest from DockerInfrastructureLiveTest
